pyproject.toml
[tool.poetry]
name = "txtclassy"
version = "0.1.0"
description = ""
authors = ["chavalee <sing@30secondstofly.com>"]

[tool.poetry.dependencies]
python = "^3.7"
ipython = "^7.13.0"
sklearn = "^0.0"
nltk = "^3.4.5"
torch = "^1.4.0"
torchtext = "^0.5.0"
pandas = "^1.0.3"
numpy = "^1.18.2"
beautifulsoup4 = "^4.9.0"
spacy = "^2.2.4"
fastapi = "^0.54.1"
uvicorn = "^0.11.3"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
